requirements.txt
astroid==1.6.6
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
backports.functools-lru-cache==1.5
configparser==3.8.1
contextlib2==0.5.5
dateutils==0.6.6
enum34==1.1.6
filelock==3.0.12
funcsigs==1.0.2
futures==3.3.0
importlib-metadata==0.19
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mock==3.0.5
more-itertools==5.0.0
nose==1.3.7
packaging==19.1
pathlib2==2.3.4
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pylint==1.9.2
pyparsing==2.4.2
pytest==4.6.5
python-dateutil==2.8.0
pytz==2019.2
scandir==1.10.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
toml==0.10.0
tox==3.13.2
virtualenv==16.7.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
